Reorganizada as funções auxiliares/de análise
